centers the comic vertically
This centers the comic vertically. Instead of editing the original code (the code that worst_face wrote that did not include the alt text) I edited the code that has the alt text in it. The reason being that the code for centering must take into account the alt text if that exists and I personally prefer the alt text.